mock python in test
